Move validate_audio_file to process_input
