+++ modules.py
@@ -2,7 +2,6 @@ import sys
 sys.path.insert(1, './modules')
 
 from chronometer import Chronometer
-from validate_audio_file import validate_audio_file
 from convert_hhmmss_to_ms import convert_hhmmss_to_ms
 from format_ms_duration import format_ms_duration
 from execution_args import execution_args
